Move VideoWriter to own file
